models/classifiers/clip_classifier.py
"""
CLIP 분류기 구현 (미래 확장용)
"""
import torch
import numpy as np
import cv2
from PIL import Image
from typing import List, Dict, Any, Optional
from ..base.base_classifier import BaseClassifier
from config.settings import CLIP_MODEL_NAME

class CLIPClassifier(BaseClassifier):
    """CLIP 기반 특징 추출 및 분류기 (미래 구현용)"""
    
    def __init__(self, **kwargs):
        super().__init__(**kwargs)
        self.model_name = kwargs.get('model_name', CLIP_MODEL_NAME)
        self.device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
        self.text_queries = kwargs.get('text_queries', [])
        
    def load_model(self) -> bool:
        """CLIP 모델 로드"""
        try:
            print(f"📥 Loading CLIP classifier: {self.model_name}")
            # TODO: CLIP 모델 로드 구현
            # from transformers import CLIPProcessor, CLIPModel
            # self.processor = CLIPProcessor.from_pretrained(self.model_name)
            # self.model = CLIPModel.from_pretrained(self.model_name)
            # self.model = self.model.to(self.device)
            
            print("⚠️  CLIP classifier is not yet implemented")
            return False
        except Exception as e:
            print(f"❌ Failed to load CLIP classifier: {e}")
            return False
    
    def extract_features(self, frame: np.ndarray, bbox: List[int]) -> Optional[np.ndarray]:
        """CLIP으로 이미지 특징 추출"""
        if not self.is_loaded():
            return None
            
        try:
            x, y, w, h = bbox
            
            # 바운딩 박스 영역 크롭
            cropped = frame[y:y+h, x:x+w]
            if cropped.size == 0:
                return None
            
            # BGR을 RGB로 변환
            cropped_rgb = cv2.cvtColor(cropped, cv2.COLOR_BGR2RGB)
            pil_image = Image.fromarray(cropped_rgb)
            
            # TODO: CLIP 특징 추출 구현
            # inputs = self.processor(images=pil_image, return_tensors="pt")
            # inputs = {k: v.to(self.device) for k, v in inputs.items()}
            # 
            # with torch.no_grad():
            #     image_features = self.model.get_image_features(**inputs)
            #     return image_features.cpu().numpy()
            
            return None
            
        except Exception as e:
            print(f"CLIP feature extraction error: {e}")
            return None
    
    def classify_features(self, features: np.ndarray, base_class: str) -> str:
        """CLIP 특징을 사용한 텍스트 기반 분류"""
        if not self.is_loaded() or features is None:
            return base_class
        
        try:
            # TODO: 텍스트 쿼리와 이미지 특징 간 유사도 계산
            # text_inputs = self.processor(text=self.text_queries, return_tensors="pt", padding=True)
            # text_inputs = {k: v.to(self.device) for k, v in text_inputs.items()}
            # 
            # with torch.no_grad():
            #     text_features = self.model.get_text_features(**text_inputs)
            #     similarities = torch.cosine_similarity(features, text_features)
            #     best_idx = similarities.argmax().item()
            #     
            #     if similarities[best_idx] > self.similarity_threshold:
            #         return self.text_queries[best_idx]
            
            return base_class
            
        except Exception as e:
            print(f"CLIP classification error: {e}")
            return base_class
    
    def set_text_queries(self, queries: List[str]) -> None:
        """분류용 텍스트 쿼리 설정"""
        self.text_queries = queries
    
    def get_model_info(self) -> Dict[str, Any]:
        """CLIP 분류기 정보 반환"""
        base_info = super().get_model_info()
        base_info.update({
            'model_name': self.model_name,
            'text_queries': self.text_queries,
            'num_queries': len(self.text_queries)
        })
        return base_info
