requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
ultralytics>=8.3.180
torch>=2.1.0
torchvision>=0.16.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
transformers>=4.35.0
requests>=2.31.0
scikit-learn>=1.3.0
timm>=1.0.19
git+https://github.com/openai/CLIP.git
ftfy>=6.0.0
regex>=2023.0.0 